Update Accrual to float

+++ internal/domain/model/accrual.go
@@ -3,5 +3,5 @@ package model
 type Accrual struct {
 	Order   string        `json:"order"`
 	Status  AccrualStatus `json:"status"`
-	Accrual *int          `json:"accrual,omitempty"`
+	Accrual *float64      `json:"accrual,omitempty"`
 }
